requirements.txt
# Dependencies
fastapi
uvicorn[standard]
sqlalchemy
pydantic
email_validator
pydantic_settings
pymysql
bcrypt
python-jose[cryptography]
python-multipart
alembic
transformers
torch
finnhub-python
pandas
yfinance
tensorflow
sklearn
SentencePiece
pandas
yfinance
scikit-learn
tensorflow
httpx
ta
prophet
